move category-modal to its own file
